# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  analytics.LogEntry:
    properties:
      api_key:
        type: string
      cost_usd:
        type: number
      customer_id:
        type: string
      input_tokens:
        type: integer
      metadata:
        additionalProperties: true
        type: object
      model:
        type: string
      output_tokens:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      team_id:
        type: string
      timestamp:
        type: string
    type: object
  analytics.SpendBucket:
    properties:
      cost_usd:
        type: number
      input_tokens:
        type: integer
      output_tokens:
        type: integer
      requests:
        type: integer
      total_tokens:
        type: integer
    type: object
  analytics.SpendLogsResponse:
    properties:
      has_more:
        type: boolean
      logs:
        items:
          $ref: '#/definitions/analytics.LogEntry'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  analytics.SpendReport:
    properties:
      by_api_key:
        additionalProperties:
          $ref: '#/definitions/analytics.SpendBucket'
        type: object
      by_customer:
        additionalProperties:
          $ref: '#/definitions/analytics.SpendBucket'
        type: object
      by_model:
        additionalProperties:
          $ref: '#/definitions/analytics.SpendBucket'
        type: object
      by_team:
        additionalProperties:
          $ref: '#/definitions/analytics.SpendBucket'
        type: object
      generated_at:
        type: string
      time_range:
        $ref: '#/definitions/analytics.TimeRange'
      total_cost_usd:
        type: number
      total_requests:
        type: integer
      total_tokens:
        $ref: '#/definitions/analytics.SpendTokens'
    type: object
  analytics.SpendTokens:
    properties:
      input:
        type: integer
      output:
        type: integer
      total:
        type: integer
    type: object
  analytics.TimeRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
  models.CacheControl:
    properties:
      type:
        description: e.g. "ephemeral"
        type: string
    type: object
  models.Choice:
    description: A single response choice.
    properties:
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/models.Message'
    type: object
...
  models.JSONSchema:
    properties:
      name:
        type: string
      schema:
        additionalProperties: {}
        type: object
      strict:
        type: boolean
    type: object
  models.LLMRequest:
    description: Unified chat completion request compatible with OpenAI, Anthropic,
      and 100+ providers.
    properties:
      frequency_penalty:
        type: number
      max_tokens:
        description: Maximum tokens to generate.
        example: 4096
        type: integer
      messages:
        description: The conversation messages.
        items:
          $ref: '#/definitions/models.Message'
        type: array
      model:
        description: The model identifier (e.g. "gpt-4o", "claude-3-opus").
        example: gpt-4o
        type: string
      presence_penalty:
        type: number
      rag_enabled:
        type: boolean
      response_format:
        allOf:
        - $ref: '#/definitions/models.ResponseFormat'
        description: ResponseFormat controls structured output.
      stop:
        description: Stop sequences.
        items:
          type: string
        type: array
      stream:
        type: boolean
      temperature:
        description: Sampling temperature (0-2).
        example: 0.7
        type: number
      tools:
        items:
          $ref: '#/definitions/models.ToolDefinition'
        type: array
      top_p:
        description: Nucleus sampling (0-1).
        example: 1
        type: number
    type: object
  models.LLMResponse:
    description: Unified LLM response.
    properties:
      choices:
        items:
          $ref: '#/definitions/models.Choice'
        type: array
      created:
        example: 1700000000
        type: integer
      id:
        example: chatcmpl-123
        type: string
      model:
        example: gpt-4o
        type: string
      object:
        example: chat.completion
        type: string
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  models.Message:
    properties:
      cache_control:
        allOf:
        - $ref: '#/definitions/models.CacheControl'
        description: |-
          CacheControl enables prompt caching (Anthropic beta feature).
          When set to {"type":"ephemeral"}, Anthropic caches this message to reduce costs.
      content:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.MessageRole'
      tool_call_id:
        type: string
      tool_calls:
        items:
          $ref: '#/definitions/models.ToolCall'
        type: array
      tool_result:
        type: string
    type: object
  models.MessageRole:
    enum:
    - system
    - user
    - assistant
    - tool
    type: string
    x-enum-varnames:
    - RoleSystem
    - RoleUser
    - RoleAssistant
    - RoleTool
  models.ResponseFormat:
    properties:
      json_schema:
        $ref: '#/definitions/models.JSONSchema'
      type:
        description: '"json_schema" | "json_object" | "text"'
        type: string
    type: object
...
  models.ToolCall:
    properties:
      function:
        $ref: '#/definitions/models.ToolFunction'
      id:
        type: string
      type:
        type: string
    type: object
  models.ToolDefinition:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        additionalProperties: true
        type: object
    type: object
  models.ToolFunction:
    properties:
      arguments:
        type: string
      name:
        type: string
    type: object
  models.Usage:
    description: Token usage statistics for a completion request.
    properties:
      completion_tokens:
        example: 50
        type: integer
      prompt_tokens:
        example: 100
        type: integer
      total_tokens:
        example: 150
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: contact@ayoubzulfiqar.com
    name: AeroLLM Contributors
    url: https://github.com/ayoubzulfiqar/aerollm
  description: AeroLLM is a high-performance, multi-provider LLM gateway that serves
    as a drop-in replacement for the LiteLLM proxy. It supports 100+ LLM providers
    with advanced routing, virtual keys, semantic caching, spend analytics, and async
    observability callbacks.
  license:
    name: MIT
    url: https://github.com/ayoubzulfiqar/aerollm/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: AeroLLM API
  version: "1.0"
paths:
...
  /global/spend/logs:
    get:
      description: Returns paginated, detailed transaction logs for a specific key
        or customer.
      parameters:
      - description: Filter by api_key, customer_id, or team_id
        in: query
        name: filter
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Entries per page (default 50, max 500)
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.SpendLogsResponse'
      summary: Get spend logs
      tags:
      - analytics
  /global/spend/report:
    get:
      description: Aggregates spend over a time range, grouped by api_key, customer,
        team, or model.
      parameters:
      - description: RFC3339 start timestamp
        in: query
        name: start
        required: true
        type: string
      - description: RFC3339 end timestamp
        in: query
        name: end
        required: true
        type: string
      - description: 'Group by: api_key, customer, team, model'
        in: query
        name: group_by
        type: string
      - description: Optional filter value
        in: query
        name: filter
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.SpendReport'
      summary: Get spend report
      tags:
      - analytics
...
  /v1/chat/completions:
    post:
      consumes:
      - application/json
      description: Send a chat completion request to an LLM provider.
      parameters:
      - description: Chat completion request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.LLMRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LLMResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat completion
      tags:
      - chat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"