# internal/api/handler.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"strings"
	"time"

	"github.com/ayoubzulfiqar/aerollm/internal/agent"
	"github.com/ayoubzulfiqar/aerollm/internal/billing"
	"github.com/ayoubzulfiqar/aerollm/internal/cache"
	"github.com/ayoubzulfiqar/aerollm/internal/analytics"
	"github.com/ayoubzulfiqar/aerollm/internal/callbacks"
	"github.com/ayoubzulfiqar/aerollm/internal/contextmgr"
	"github.com/ayoubzulfiqar/aerollm/internal/finops"
	"github.com/ayoubzulfiqar/aerollm/internal/ledger"
	"github.com/ayoubzulfiqar/aerollm/internal/models"
	"github.com/ayoubzulfiqar/aerollm/internal/providers"
	"github.com/ayoubzulfiqar/aerollm/internal/ratelimit"
	"github.com/ayoubzulfiqar/aerollm/internal/router"
	"github.com/ayoubzulfiqar/aerollm/pkg/telemetry"
)

// LoggerInterface defines the logging interface.
type LoggerInterface interface {
	Info(msg string, keysAndValues ...interface{})
	Error(msg string, keysAndValues ...interface{})
}

// Handler handles HTTP requests for the LLM gateway.
type Handler struct {
	Router      *router.Router
	Agent       *agent.AgentEngine
	Cache       *cache.RedisCache
	RateLimiter ratelimit.RateLimiter
	Telemetry   *telemetry.Provider
	Logger      LoggerInterface

	// CallbackMgr dispatches observability callbacks asynchronously.
	CallbackMgr *callbacks.CallbackManager
	// SemanticCache provides production-grade semantic caching via embeddings.
	SemanticCache *cache.VectorSemanticCache
	// Analytics records spend data for /global/spend/report and /global/spend/logs.
	Analytics *analytics.AnalyticsEngine

	Advanced interface {
		ResumeApproval(ctx context.Context, approvalID string, approved bool, req *models.LLMRequest) (*models.LLMResponse, error)
	}
	UsageRecorder *finops.CostTracker
	BudgetChecker interface {
		CheckBudget(ctx context.Context, apiKey string, estimatedCost float64) (float64, error)
	}
	Ledger interface {
		Append(ctx context.Context, record ledger.LedgerRecord) error
		Latest(ctx context.Context) (*ledger.LedgerRecord, error)
	}
	UsageSink chan<- billing.MeterEntry

	// ModelResolver resolves a provider for a given model alias. Optional.
	ModelResolver func(model string) (providers.Provider, bool)
}

// NewHandler creates a new Handler with dependency injection.
func NewHandler(router *router.Router, agent *agent.AgentEngine, cache *cache.RedisCache, rl ratelimit.RateLimiter, tp *telemetry.Provider, logger LoggerInterface) *Handler {
	return &Handler{
		Router:      router,
		Agent:       agent,
		Cache:       cache,
		RateLimiter: rl,
		Telemetry:   tp,
		Logger:      logger,
	}
}

// ChatCompletions handles the /v1/chat/completions endpoint.
// @Summary Chat completion
// @Description Send a chat completion request to an LLM provider.
// @Tags chat
// @Accept json
// @Produce json
// @Param req body models.LLMRequest true "Chat completion request"
// @Success 200 {object} models.LLMResponse
// @Failure 400 {object} map[string]string
// @Router /v1/chat/completions [post]
func (h *Handler) ChatCompletions(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	start := time.Now()
	ctx, span := h.Telemetry.StartSpan(ctx, "ChatCompletions")
	defer span.End()

	var req models.LLMRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		h.Logger.Error("invalid request", "error", err)
		http.Error(w, `{"error":"invalid request"}`, http.StatusBadRequest)
		return
	}
	if req.Model == "" {
		h.Logger.Error("missing model")
		http.Error(w, `{"error":"missing model"}`, http.StatusBadRequest)
		return
	}

	apiKey := r.Header.Get("Authorization")
	if h.UsageRecorder != nil && apiKey != "" {
		_ = h.UsageRecorder.RecordUsage(ctx, finops.CostRequest{APIKey: apiKey, Model: req.Model, Usage: &models.Usage{}})
	}

	// Cache exact-match check
	if h.Cache != nil {
		cacheKey := cache.KeyForRequest(&req)
		if cached, err := h.Cache.GetExact(cacheKey); err == nil && cached != nil {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)
			w.Write(cached.Response)
			telemetry.RecordCacheHit(true)
			h.Logger.Info("cache hit", "key", cacheKey)
			return
		}
	}

	// Semantic cache check — uses embedding-based cosine similarity.
	if h.SemanticCache != nil {
		semKey := cache.KeyForRequest(&req)
		// Serialize the request prompt for embedding.
		promptText := req.Model
		for _, msg := range req.Messages {
			if msg.Content != nil {
				promptText += " " + *msg.Content
			}
		}
		if hit, err := h.SemanticCache.Search(ctx, promptText); err == nil && hit != nil {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)
			w.Write(hit.Response)
			telemetry.RecordCacheHit(true)
			h.Logger.Info("semantic cache hit", "key", semKey)
			return
		}
	}

	// Auto-summarization to keep context within model limits
	if len(req.Messages) > 0 {
		ctxMgr := contextmgr.NewContextManager(map[string]int{
			req.Model: 4096,
		})
		_, _ = ctxMgr.MaybeSummarize(ctx, req.Model, req.Messages)
	}

	// Rate limiting
	if h.RateLimiter != nil {
		allowed, err := h.RateLimiter.Allow(ctx, apiKey, req.Model)
		if err != nil || !allowed {
			h.Logger.Error("rate limited", "error", err)
			http.Error(w, `{"error":"rate limited"}`, http.StatusTooManyRequests)
			return
		}
	}

	// Budget pre-check
	if h.BudgetChecker != nil && apiKey != "" {
		if _, err := h.BudgetChecker.CheckBudget(ctx, apiKey, 0); err != nil {
			h.Logger.Error("budget exceeded", "error", err)
			http.Error(w, `{"error":"budget exceeded"}`, http.StatusPaymentRequired)
			return
		}
	}

	// Route to provider
	selectedProvider, err := h.Router.Route(ctx, &req)
	if err != nil {
		h.Logger.Error("routing error", "error", err)
		http.Error(w, `{"error":"routing failed"}`, http.StatusInternalServerError)
		telemetry.RecordError()
		return
	}

	// Provider-aware agentic tool execution loop
	agentWithProvider := &agent.AgentEngine{
		Provider:      h.Agent.Provider,
		MaxIterations: h.Agent.MaxIterations,
		ToolTimeout:   h.Agent.ToolTimeout,
		MaxConcurrent: h.Agent.MaxConcurrent,
		ToolCache:     h.Agent.ToolCache,
		Registry:      h.Agent.Registry,
	}
	resp, err := agentWithProvider.RunToolExecutionLoop(ctx, &req)
	if err != nil {
		h.Logger.Error("agent error", "error", err)
		http.Error(w, `{"error":"agent execution failed"}`, http.StatusInternalServerError)
		telemetry.RecordError()
		return
	}
	if resp == nil {
		h.Logger.Error("empty agent response")
		http.Error(w, `{"error":"empty agent response"}`, http.StatusInternalServerError)
		return
	}
	if selectedProvider != nil {
		resp.Model = selectedProvider.Name()
	}

	respBytes, err := json.Marshal(resp)
	if err != nil {
		h.Logger.Error("encoding error", "error", err)
		http.Error(w, `{"error":"encoding failed"}`, http.StatusInternalServerError)
		return
	}

	// Cache exact-match store
	if h.Cache != nil {
		cacheKey := cache.KeyForRequest(&req)
		_ = h.Cache.SetExact(cacheKey, respBytes, 0)
	}

	// Record analytics for spend reporting.
	if h.Analytics != nil && selectedProvider != nil {
		providerName := selectedProvider.Name()
		apiKey := extractAPIKey(r.Header.Get("Authorization"))
		h.Analytics.RecordFromUsage(resp.ID, apiKey, "", "", req.Model, providerName, resp.Usage, 0)
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	_ = json.NewEncoder(w).Encode(resp)

	telemetry.RecordRequestCount(selectedProvider.Name(), 1)
	telemetry.RecordLatencyMs(float64(time.Since(start).Milliseconds()))

	// Store in semantic cache for future similarity-based lookups.
	if h.SemanticCache != nil {
		semKey := cache.KeyForRequest(&req)
		promptText := req.Model
		for _, msg := range req.Messages {
			if msg.Content != nil {
				promptText += " " + *msg.Content
			}
		}
		_ = h.SemanticCache.Upsert(ctx, semKey, promptText, respBytes, map[string]interface{}{
			"model": req.Model,
		})
	}

	// Fire async callbacks — response is already sent to client.
	if h.CallbackMgr != nil && resp.Usage != nil {
		// Build token counts from the response usage.
		tokenCount := map[string]int{}
		if resp.Usage.PromptTokens > 0 {
			tokenCount["input"] = resp.Usage.PromptTokens
		}
		if resp.Usage.CompletionTokens > 0 {
			tokenCount["output"] = resp.Usage.CompletionTokens
		}

		// Convert messages to interface{} for callback payload.
		msgMaps := make([]map[string]interface{}, 0, len(req.Messages))
		for _, m := range req.Messages {
			msgMap := map[string]interface{}{
				"role": string(m.Role),
			}
			if m.Content != nil {
				msgMap["content"] = *m.Content
			}
			msgMaps = append(msgMaps, msgMap)
		}

		reqData := &callbacks.CallbackRequestData{
			RequestID:  fmt.Sprintf("req-%d", time.Now().UnixNano()),
			Model:      req.Model,
			Provider:   selectedProvider.Name(),
			Messages:   msgMaps,
			CostUSD:    float64(0), // costTracker could be queried here
			Timestamp:  time.Now(),
		}
		respData := &callbacks.CallbackResponseData{
			ResponseID:   resp.ID,
			Usage:        map[string]interface{}{"prompt_tokens": resp.Usage.PromptTokens, "completion_tokens": resp.Usage.CompletionTokens},
			Model:        resp.Model,
			Provider:     selectedProvider.Name(),
			LatencyMs:    time.Since(start).Milliseconds(),
			TokenCount:   tokenCount,
			Metadata:     map[string]interface{}{"stream": req.Stream},
			Timestamp:    time.Now(),
		}
		h.CallbackMgr.FireSuccess(reqData, respData)
	}

	if h.Ledger != nil {
		reqBytes, _ := json.Marshal(req)
		respBytes, _ := json.Marshal(resp)
		latest, _ := h.Ledger.Latest(ctx)
		prevHash := ""
		if latest != nil {
			prevHash = latest.ChainHash
		}
		chainHash := ledger.ComputeChainHash(prevHash, string(reqBytes), string(respBytes))
		_ = h.Ledger.Append(ctx, ledger.LedgerRecord{
			Timestamp:         time.Now().UTC(),
			PrevHash:          prevHash,
			RequestPayload:    string(reqBytes),
			ResponsePayload:   string(respBytes),
			ChainHash:         chainHash,
		})
	}
}

// ResumeApproval handles the /v1/agents/approvals/{id} endpoint.
func (h *Handler) ResumeApproval(w http.ResponseWriter, r *http.Request) {
	if h.Advanced == nil {
		http.Error(w, `{"error":"advanced agent not enabled"}`, http.StatusNotImplemented)
		return
	}

	approvalID := strings.TrimPrefix(r.URL.Path, "/v1/agents/approvals/")
	if approvalID == "" {
		http.Error(w, `{"error":"missing approval id"}`, http.StatusBadRequest)
		return
	}

	var req struct {
		Approved bool `json:"approved"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, `{"error":"invalid request"}`, http.StatusBadRequest)
		return
	}

	resp, err := h.Advanced.ResumeApproval(r.Context(), approvalID, req.Approved, &models.LLMRequest{})
	if err != nil {
		h.Logger.Error("resume approval failed", "error", err)
		http.Error(w, `{"error":"approval failed"}`, http.StatusBadRequest)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}

// ProviderHealth represents provider health status for HTTP responses.
type ProviderHealth struct {
	Name        string `json:"name"`
	Type        string `json:"type"`
	Healthy     bool   `json:"healthy"`
	CircuitOpen bool   `json:"circuit_open"`
}

// HealthProviders returns the health status of all registered providers.
func (h *Handler) HealthProviders() []ProviderHealth {
	var out []ProviderHealth
	if h.Router == nil {
		return out
	}
	for _, cb := range h.Router.Providers() {
		health := cb.Health()
		out = append(out, ProviderHealth{
			Name:        health.Name,
			Type:        string(health.Type),
			Healthy:     health.Healthy,
			CircuitOpen: health.CircuitOpen,
		})
	}
	return out
}

// HealthCheck handles the /health endpoint.
func (h *Handler) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(map[string]string{"status": "ok"})
}

// ReadyCheck handles the /ready endpoint.
func (h *Handler) ReadyCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(map[string]string{"ready": "true"})
}

// AdvancedLoopHandler wires the advanced execution loop into the HTTP handler.
type AdvancedLoopHandler struct {
	Agent  *agent.AgentEngine
	Logger LoggerInterface
	Hooks  []agent.LoopHook
}

// NewAdvancedLoopHandler creates a new handler wrapper.
func NewAdvancedLoopHandler(a *agent.AgentEngine, logger LoggerInterface, hooks []agent.LoopHook) *AdvancedLoopHandler {
	return &AdvancedLoopHandler{Agent: a, Logger: logger, Hooks: hooks}
}

// Run executes the advanced agent loop for a request.
func (h *AdvancedLoopHandler) Run(ctx context.Context, req *models.LLMRequest) (*models.LLMResponse, error) {
	if h.Agent == nil {
		return nil, nil
	}
	return h.Agent.RunAdvancedExecutionLoop(ctx, req, agent.AdvancedLoopOptions{Hooks: h.Hooks})
}

// RateLimitMiddleware wraps a handler with rate limiting.
func RateLimitMiddleware(next http.HandlerFunc, rl ratelimit.RateLimiter) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		next(w, r)
	}
}

// extractAPIKey extracts the API key from the Authorization header.
func extractAPIKey(auth string) string {
	if auth == "" {
		return ""
	}
	if len(auth) > 7 && auth[:7] == "Bearer " {
		return auth[7:]
	}
	return auth
}